Removed clippy error with bundle_provider, put engine_provider in core
