+++ crates/cli/Cargo.toml
@@ -17,7 +17,6 @@ path = "src/main.rs"
 contender_core = { workspace = true }
 contender_sqlite = { workspace = true }
 contender_testfile = { workspace = true }
-contender_engine_provider = { workspace = true }
 
 ansi_term = { workspace = true }
 serde = { workspace = true }
