+++ crates/core/src/engine_provider/auth_provider_op.rs
@@ -1,4 +1,4 @@
-use crate::{util::read_jwt_file, AdvanceChain};
+use crate::engine_provider::{util::read_jwt_file, AdvanceChain};
 use alloy::{
     eips::Encodable2718,
     hex as make_hex,
