+++ crates/core/src/engine_provider/util.rs
