+++ crates/core/src/test_scenario.rs
@@ -1,6 +1,9 @@
 use crate::agent_controller::AgentStore;
 use crate::buckets::Bucket;
+use crate::bundle_provider::new_basic_bundle;
+use crate::bundle_provider::BundleClient;
 use crate::db::{DbOps, NamedTx};
+use crate::engine_provider::AdvanceChain;
 use crate::error::ContenderError;
 use crate::generator::named_txs::ExecutionRequest;
 use crate::generator::templater::Templater;
@@ -25,9 +28,6 @@ use alloy::rpc::types::TransactionRequest;
 use alloy::serde::WithOtherFields;
 use alloy::signers::local::{LocalSigner, PrivateKeySigner};
 use alloy::transports::http::reqwest::Url;
-use crate::bundle_provider::new_basic_bundle;
-use crate::bundle_provider::BundleClient;
-use contender_engine_provider::AdvanceChain;
 use futures::{Stream, StreamExt};
 use std::collections::{BTreeMap, HashMap};
 use std::ops::Deref;
