--- crates/engine_provider/CHANGELOG.md
@@ -1,15 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
-## [0.1.2](https://github.com/flashbots/contender/releases/tag/contender_engine_provider-v0.1.2) - 2025-05-14
-
-### Other
-
-- ci publish ([#215](https://github.com/flashbots/contender/pull/215))
-- engine_ calls to advance chain manually ([#165](https://github.com/flashbots/contender/pull/165))
