--- crates/engine_provider/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "contender_engine_provider"
-version = "0.1.2"
-edition.workspace = true
-rust-version.workspace = true
-authors.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-description = "Contender engine_ API provider"
-
-[lib]
-name = "contender_engine_provider"
-path = "src/lib.rs"
-
-[dependencies]
-alloy = { workspace = true, features = [
-    "full",
-    "node-bindings",
-    "rpc-types-mev",
-    "json-rpc",
-    "provider-engine-api",
-] }
-alloy-rpc-types-engine = { workspace = true, features = ["std", "jwt"] }
-alloy-chains = { version = "0.1.64", default-features = false }
-alloy-json-rpc = { workspace = true }
-tracing = { workspace = true }
-futures = { workspace = true }
-async-trait.workspace = true
-tokio = { workspace = true }
-thiserror = { workspace = true }
-tower = { workspace = true }
-eyre = { workspace = true }
-
-reth-node-api = { git = "https://github.com/paradigmxyz/reth", tag = "v1.3.4", default-features = false }
-reth-optimism-node = { git = "https://github.com/paradigmxyz/reth", tag = "v1.3.4" }
-op-rbuilder = { git = "https://github.com/flashbots/rbuilder", branch = "brock/tester-engine-api", default-features = false }
-op-alloy-consensus = { version = "0.11.2", default-features = false }
