--- crates/engine_provider/README.md
@@ -1,5 +1,0 @@
-# this crate was stolen
-
-most of the credit goes to:
-- [reth-bench](https://github.com/paradigmxyz/reth/tree/main/bin/reth-bench); @rjected and reth contributors.
-- [op-rbuilder tester](https://github.com/flashbots/rbuilder/tree/develop/crates/op-rbuilder/src/tester); @ferranbt & rbuilder contributors
